+++ docs/changelog.rst
@@ -6,8 +6,38 @@ Changelog
 #########
 
 ******************
-0.1.3 (unreleased)
+0.1.4 (unreleased)
 ******************
+
+******************
+0.1.3 (2022-05-05)
+******************
+
+New Features
+============
+- Append the Abaqus journal Builder managed targets automatically (:issue:`18`, :merge:`10`). By `Kyle Brindley`_.
+- Separate the common custom builders from the EABM SCons project definition (:issue:`19`, :merge:`11`). By `Kyle
+  Brindley`_.
+- Add a variable to pass through additional Abaqus command line arguments to the Abaqus journal file builder
+  (:issue:`19`, :merge:`11`). By `Kyle Brindley`_.
+- Add the SCons target definition equivalent to the ECMF and CMake-simulation Abaqus simulation execution (:issue:`21`,
+  :merge:`13`). By `Kyle Brindley`_.
+
+Bug fixes
+=========
+- Avoid modifying the contents or timestamp of input files in Abaqus journal files (:issue:`17`, :merge:`12`). By `Kyle
+  Brindley`_.
+
+Documentation
+=============
+- Add SCons custom builder documentation for the build system (:issue:`19`, :merge:`11`). By `Kyle Brindley`_.
+- Separate the Scons build system documentation from the associated SCons-EABM documentation (:issue:`26`, :merge:`16`).
+  By `Kyle Brindley`_.
+
+Internal Changes
+================
+- Remove the dummy ``{job_name}.touch`` file from the Abaqus wrapper. SCons does not automatically delete target file(s)
+  when the build fails like GNU Make or CMake does (:issue:`24`, :merge:`14`). By `Kyle Brindley`_.
 
 ******************
 0.1.2 (2022-05-04)